+++ src/main.ts
@@ -19,7 +19,7 @@ import { makeInvoke } from "./sidecar/bridge.js";
 import { LaneReader } from "./sidecar/lanes.js";
 import { UsesReader } from "./sidecar/uses.js";
 
-export const version = "0.2.0";
+export const version = "0.2.1";
 
 export interface Started {
   app: FastifyInstance;